+++ src/Badger/UserBundle/DependencyInjection/UserExtension.php
@@ -24,5 +24,6 @@ class UserExtension extends Extension
         $loader->load('repositories.yml');
         $loader->load('savers.yml');
         $loader->load('services.yml');
+        $loader->load('serializers.yml');
     }
 }