+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Added
+- Add search functionality on users. You can now search users by their usernames!
+
 ## [0.5.2] - 2016-06-09
 ### Fixed
 - Fix fatal error on new user registration